Refactoring ROI items into silx.gui.plot.items
